edit totalmente funcional. TODO: carregar images no view e mensagens

+++ app/controllers/CrudProductController.php
@@ -44,9 +44,6 @@ class CrudProductController
         $this->queryBuilder->table("products")->insert([$_POST['name'], $_POST['price'], $_POST['selection-category'], $_POST['description']]);
         $idProduct = $this->queryBuilder->table("products")->select("*")->where("name", "=", $_POST['name'])->commit()[0]['id'];
 
-        //Apenas para editar
-        // $prepare = $this->queryBuilder->table("images")->delete()->where('productId', '=', $idProduct);
-        // $prepare->commit();
 
         $allowedFiles = ['jpg', 'png', 'jpeg'];
 
@@ -75,22 +72,22 @@ class CrudProductController
 
 
 
-    public function update()
-    {
+    public function update(){
         
         $prepare = $this->queryBuilder->table("products")->update(['name', 'price', 'categoryId', 'description'], [$_POST['name'], $_POST['price'], $_POST['selection'], $_POST['description']])->where('id', '=', $_POST['id']);
         $prepare->commit();
         $idProduct = $this->queryBuilder->table("products")->select("*")->where("name", "=", $_POST['name'])->commit()[0]['id'];
 
-        //Apenas para editar
-        $prepare = $this->queryBuilder->table("images")->delete()->where('productId', '=', $idProduct);
-        $prepare->commit();
-
         $allowedFiles = ['jpg', 'png', 'jpeg'];
-
+       
         $files = $_FILES['files'];
-
         $names = $files['name'];
+
+        if($_FILES['files']['error'][0] != 4 && $_FILES['files']['size'][0] != 0): //!Verifica se os arquivos foram alterados, se sim deleta os anteriores
+            $prepare = $this->queryBuilder->table("images")->delete()->where('productId', '=', $idProduct);
+            $prepare->commit();
+        endif;
+
         for ($i = 0; $i < count($names); $i++) :
             $extension = explode('.', $names[$i]);
             $extension = end($extension);
@@ -101,7 +98,6 @@ class CrudProductController
                 move_uploaded_file($files['tmp_name'][$i], $destino);
             endif;
         endfor;
-
 
         header("Location: /admin/products");
     }
